requirements.txt
Flask>=3.0.0,<4
requests>=2.32.0
python-dotenv>=1.0.1
nltk>=3.9.0
numpy>=2.1.0
pandas>=2.2.2
matplotlib>=3.9.0
seaborn>=0.13.0
